+++ packages/brick_offline_first_with_graphql/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Update default of offline queue from 0 seconds to 5 seconds
+* Notify SQLite subscribers when a remote update has occurred.
 
 ## 1.3.0